+++ src/app/airlist/airlist.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-airlist',
+  templateUrl: './airlist.component.html',
+  styleUrls: ['./airlist.component.css']
+})
+export class AirlistComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit(): void {
+  }
+
+}