+++ src/app/hotel/hotel.component.ts
@@ -1,4 +1,7 @@
 import { Component, OnInit } from '@angular/core';
+import { ReservtwoComponent } from '../reservtwo/reservtwo.component';
+import { ReservthreeComponent } from '../reservthree/reservthree.component';
+import { MatDialog} from '@angular/material/dialog';
 
 @Component({
   selector: 'app-hotel',
@@ -7,8 +10,13 @@ import { Component, OnInit } from '@angular/core';
 })
 export class HotelComponent implements OnInit {
 
-  constructor() { }
-
+  constructor(public dialog: MatDialog) { }
+  openDialog1(){
+    this.dialog.open(ReservtwoComponent);
+  }
+  openDialog2(){
+    this.dialog.open(ReservthreeComponent);
+  }
   ngOnInit(): void {
   }