+++ src/app/idcard/idcard.component.ts
@@ -1,4 +1,6 @@
 import { Component, OnInit } from '@angular/core';
+import { ReservoneComponent } from '../reservone/reservone.component';
+import { MatDialog} from '@angular/material/dialog';
 
 @Component({
   selector: 'app-idcard',
@@ -7,8 +9,10 @@ import { Component, OnInit } from '@angular/core';
 })
 export class IdcardComponent implements OnInit {
 
-  constructor() { }
-
+  constructor(public dialog: MatDialog) { }
+  openDialog(){
+    this.dialog.open(ReservoneComponent);
+  }
   ngOnInit(): void {
   }