+++ src/app/reservone/reservone.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-reservone',
+  templateUrl: './reservone.component.html',
+  styleUrls: ['./reservone.component.css']
+})
+export class ReservoneComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit(): void {
+  }
+
+}