+++ src/app/reservthree/reservthree.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-reservthree',
+  templateUrl: './reservthree.component.html',
+  styleUrls: ['./reservthree.component.css']
+})
+export class ReservthreeComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit(): void {
+  }
+
+}