+++ src/app/reservtwo/reservtwo.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-reservtwo',
+  templateUrl: './reservtwo.component.html',
+  styleUrls: ['./reservtwo.component.css']
+})
+export class ReservtwoComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit(): void {
+  }
+
+}